libipsecconf: support old-style ",," to mean "\," in specifying leftid/rightid for ID_DER_ASN1_DN
libreswan 3.30 and earlier allowed one to use ,, to mask in an RDN part.

For example: leftid="CN=foo OU=Global,, Services, C=CA"

The current code no longer sees this special case when comparing. It does support
using \, to mask the comma. And this is compatible with sending it into NSS.

So let the parser change ",," to "\,"