FIPS: In lsw_get_fips_mode() ensure NSS_IsInitialized()
While there are now no callers of lsw_get_fips_mode() or
libreswan_fipsmode() that call it without initializing an
NSS databse, this assertion ensures in the future this is
not introduced either.

Without NSS being initialized, PK11_IsFIPS() does not return
a valid result and thus should not be used in such a way.