testing: extend ikev2-05-basic-psk-oneconf-namespaces to test updown traffic counters
Test for PLUTO_INBYTES to ensure traffic counters get to updown script.
This was broken in libreswan 3.30-3.32

The test is a little fragile, as the PLUTO_INBYTES string could get wrapped
when other parameters to updown change. If this becomes in issue in the
future, a custom updown script could be used that performs the test.