testing: added testing/utils/namespace-runner.py
Usage: namespace-runner.py [-j num] [---verbose] [--diff] testname1 testname2 ...

Options:
  -h, --help     show this help message and exit
  -j NUM         number of tests to run in parallel
  -v, --verbose  show verbose nsrun output - output will interleave with
                 threads
  -d, --diff     show diff output after all tests have run

This allows for a selection of tests to be run with namespaces, and be
fired of in parallel. All diffs can be output to the screen so it can
be run and the data stored without needing to keep the OUTPUT directories.

This can be used for Fedora CI gating to wrap our upstream tests