pluto: add co_serial_t serial_from to struct connection.
This can be used in cases where we need to store the template
connection from which the instance was instantiated.

This will be used for session resumption. An instance that is
going to be suspended, will list the serial_from in the session
cookie. Upon session resumption, the cookie data can be used to
instantiate a new connection from the original template.

If the template has since reloaded, it will have a different
serial_from, and the session cookie will properly fail.

The ipsec status output now lists the serialno and serial_from
of the connection.