testing: nsrun takes a new argument --exitcode
If set, if will return 1 when a test fails. The reason this is not
done per default is because it is also used with a make check
target where it runs all the tests, and it would abort too soon then.

This option is used by testing/util/namespace-runner.py