libipsecconf: promote ah= / esp= as desired keywords over phase2alg=
This is a reversal of what we tried to do in the past. Since IKEv2
does not really talk about phase2 anymore, this term is no longer
favoured.

Ideally, phase2=ah|esp would also get renamed, but what word to use?

(type is already used for tunnel|transport, and mode= would be confused
 with transport|tunnel mode. And encapsulation=ah would be weird because
 there is no encapsulation. And no one wants ah=yes)