IKEv2: redirect: enable load-balancing feature
Add option to specify multiple destinations for redirection during
IKE_SA_INIT. In addition to that, simplify redirect code a bit.

D. Hugh Redelmeier assisted in improving this patch.

This addition enables to have one 'front' server that will be used as a
load balancer, redirecting incoming clients to a list of other servers.

Before it was possible only to have something like 'global-redirect-to=1.1.1.1'
Now one can have, i.e. 'global-redirect-to=1.1.1.1,2.2.2.2,3.3.3.3'

Signed-off-by: D. Hugh Redelmeier <hugh@mimosa.com>
Signed-off-by: Paul Wouters <pwouters@redhat.com>