pluto: remove ikeport/nat-ikeport, add listen-udp/listen-tcp
Moving UDP 500 to another port does not work anyway, because then
it would not use the non-ESP marker. Moving the global nat-t port
also makes less sense, since we can add them per connection now anyway
with leftikeport/rightikeport.

What does make sense is an option to enable/disable the global listen
ports for UDP or TCP.

This also frees up the --ikeport for whack to be used for the leftikeport/
rightikeport options.

In "ipsec status", remove the ikeport= output, and in the interface list,
add UDP or TCP to indicate which protocol it is listening for.

Note: It seems the nat_traversal.c code sometimes compared against
pluto_nat_port, which won't work when custom ikeports are used. This
will need fixing.