libipsecconf: Fix config handling of policy-label
It appears that 998aa5f18bad901d6f5c834494d06ee405053b59 changed
policy_label from being an "option" to a "string", but forgot to change
the conditional for (optionally) sending the value to pluto.
Using 3.32 this results in addconn parsing the configuration file but
simply **never** sends this value to pluto.
This breaks Labeled IPsec support.

Signed-off-by: Paul Wouters <pwouters@redhat.com>