XFRM: Add ipsec0 support by mapping it onto if_d 16384
This requires ipsec-interface= to take 0 as valid argument,
so we use UINT32_MAX to mean "no" in the parser.

When setting up the XFRM interface, if we encounter if_id 0,
we map it to 16384, but once we make a name we will use "ipsec0"
instead of "ipsec16384".

whack support was also still missing, so whack support was added
with the same mapping.

This is documented in a man page update.