connections: in add_connection() bake "connection \"%s\"": prefix into the logger
This way nested calls can assume that the correct prefix will be
automatically added.  Drop no-longer-used connection_name parameter.

For instance.  While a connection is being added:

  002 connection "distraction": bad left CA string 'Something bogus': unknown OID in ID_DER_ASN1_DN (ignored)
  036 connection "distraction": failed to add connection: leftauth= and rightauth= require ikev2

and once created it switches to the shorter:

  "distraction": ...

(the reason for two prefixes is code evolution and, I suspect, a way
to differentiate a partially constructed vs complete connection?).