testing: update certoe-17-asymmetric-cert-nat-testing-server-crash
This was a broken test. It tests a crashed server east, but east is
on OE clear-or-private, so it happilly receives plaintext from the %pass
on road that happened when it was down. And it won't trigger a tunnel.
We should copy this test case in a variant where east is also
private-or-clear to see if it replaces the %pass on road.