pluto: new config setup option ikev1-policy=<accept|drop|reject>
This option allows pluto to be configured to reject all IKEv1 packets.
If set to reject, it will return INVALID_MAJOR_VERSION. If set to drop,
it will quietly drop the packet.

If set to a value of drop or reject, it will also reject loading ikev1
connections, as these would never be able to receive any packets for
processing.