pluto: in show_kernel_alg_connection() don't use IKEv1 function.
This code is also called for IKEv2, so should not call ikev1_quick_pfs()

This is to show a different strong for PFS between parent and child.

This isn't supported anyway, so I am not sure this code was ever called
as intended. For IKEv2 it will not print "<Phase1>" without calling
ikev1_quick_pfs() that always resulted in NULL and then printing "<Phase1>"
anyways.