pluto: fixup IKEv2 pending issues - follow up of d491a68dbfb8be
While the commit split the use of the ikev1 only macro in two code
paths, with its original meaning, it showed up as ikev2 failing to
add to the pending queue.

This commit also pushes ikev2 IPsec SA's into the pending queue
if the IKE SA is not yet established.