impair: add --impair drop-{incoming,outgoing}:<message-nr>
Where <message-nr> is the Nth unique message for the given direction
(duplicates have the same number).

Once the impair has been actioned, it is discarded.  Hence, the
impairment:

	--impair drop-incoming:4

will match and drop the first arrival of the 4th message but will then
allow retransmits of that same message that message to pass through.