testing: removed-iface-01: rename 8-final.sh to 8-east-west-final.sh
When eastinit.sh is not there, testing glob()s the scripts, but looks
for the east/west/road/north/nic hostnames to see which is run where.

Where as final.sh used to be called on all hosts, final.sh (or 8-final.sh)
would not. Calling it 8-east-west-final.sh will run it as expected.