pluto: add auto=keep option and revival for instances (via co_serialno)
Problem scenario: a peer behind NAT initiates to libreswan and brings up
a tunnel. Once idle, the tunnels is torn down, but libreswan wants the
tunnel to always be up. It cannot do a regular initiate of the connection
because the peer is behind NAT so IKE traffic to port 500/4500 will fail.

auto=keep will add POLICY_UP to the auto=add, so once the instance comes up
and receives a delete, it will update its ike port to the last known port
in use, and immediately re-initiate back before the NAT router closes the
NAT mapping and the machine behind NAT becomes unreachable.

If keyingtries=0, set it to 2 so in case the machine behind NAT is gone,
the revival process ends and the instance connection is deleted.