pluto: add optional --leave-state to ipsec whack --shutdown that leaves kernel state
If something restarts pluto but wants to limit downtime as much as possible,
this can be used so that as long as the kernel still processes packets, IPsec
also keeps working.

Note: this needs some coordination with the init system. At the moment, systemd
will flush the kernel anyway in ExecStopPost