ip: restore debug log in endpoint_eq()
Add the comment:

XXX: note the <<#if 0 pendpoint()>> sprinkled all over this file.

There is (was?) code lurking in pluto that does not initialize the
ip_endpoint's .iproto field.  For instance by assigning an ip_address
to an ip_endpoint (the two are still compatible).

This dbg() line is one step towards tracking these cases down.

(If the intent is for some sort of wildcard match then either
ip_selector or ip_subnet can be used.)