pluto: Use NULL when refering to pointers
The code block here defines an array of char* pointers, but then sets
the first element to '\0', which is a char an not a pointer. Set it to
NULL instead.
This fixes the following warning from clang:

programs/pluto/ikev1_spdb_struct.c:1278:18: error: expression which evaluates to zero treated as a null pointer constant of type 'char *' [-Werror,-Wnon-literal-null-conversion]
                                                nbyte[0] = '\0';
                                                           ^~~~