pluto: change default IKE SA lifetime from 1h to 8h.
With IKEv1, the IKE SA could expire while retaining the IPsec SA.
Which meant an IKE SA of 1h would not affect an IPsec SA of 8h.

With IKEv2, if the IKE SA expires, it takes down all IPsec SA's
as well.

As a result, the same default for ikelifetime= causes very different
behaviour between IKEv1 and IKEv2 when rekey=no is set.

While it is possible with libreswan 4.x to set rekey=yes on the server
side, to ensure the connection stays up, often clients (eg Windows)
do not like it when the server initiates a rekey to them.

This was reported at various places. Some examples:

Resolves: https://github.com/libreswan/libreswan/issues/405
Resolves: https://github.com/hwdsl2/setup-ipsec-vpn/issues/913
Resolves: https://github.com/libreswan/libreswan/issues/362