Use close_range to close fds
On some systems, the file descriptor table can be large (524,288
entries on the system I’m writing this on), and the fd-closing loop in
plutomain takes a long time (2-3 *minutes*).

Linux 5.10 introduced the close_range syscall which can close huge
ranges of fds very quickly. This patch enables pluto to use this, when
it’s available.

The Makefile snippet changes detect support for close_range at build
time, using linux/close_range.h as a proxy (if it’s present, the
syscall number will also be present). At runtime, the syscall is
attempted in all cases if pluto was built with it; older kernels will
fail with ENOSYS. In any case, if close_range fails, we fall back to
the existing fd-closing behaviour, which is preserved anyway because
we want to loop over the fds from 0 to ctl_fd (excluded).

Signed-off-by: Stephen Kitt <skitt@redhat.com>