testing: use LD_PRELOAD and not the faketime wrapper when starting pluto
After:

  pluto: don't try to close all open files when only 0,1,2,lockfile,ctl_fd are open

the command:

  faketime ipsec pluto

was hanging because the test was relying on pluto closing the FDs:

  $ faketime 'tomorrow' ls -l /proc/self/fd
  lrwx------. 1 cagney cagney 64 Jan 28 20:45 3 -> /dev/shm/faketime_shm_2390789
  l-wx------. 1 cagney cagney 64 Jan 28 20:45 4 -> 'pipe:[14588018]'
  lrwx------. 1 cagney cagney 64 Jan 28 20:45 6 -> 'socket:[14579955]'

during the detach (the've got something to do with keeping all the
child processes fake times in sync).  Since the're being left open
the root faketime process didn't see a need to exit.

Change the tests to use LD_PRELOAD et.al., and put pluto at the
root of the faketime process tree.