pluto: get_newer_sa_from_connection() should not have numeric assumption
It would check serialno and assume a larger number meant a more recent
state, but this is not always true. An initiating state #1 can end up
established after a responder state #2 finished first, in which #1 is
the latest, since it will be the last one that updated c->newest_isakmp_sa

Basically, whatever is in c->newest_isakmp_sa is per definition the
newest.