kernel: pass WHY into bare_shunt_ptr()
Why not (well actuall so that debug logs can provide information
on why the lookup is happening).

Also sprinkle #if 0 pexpect() that an end's .client, .protocol and
.port are consistent.  Now that a client has been changed to a
ip_selector it's big enough to hold both the protocol and the port
(previously it was a subnet that strangely sometimes contained a
port).