pluto: Rework Labeled IPsec to send the right narrowed labels
We need to send the right labels over IKE and to the kernel XFRM stack,
otherwise we keep getting acquired for the wide %trap policy

For IKEv2, it is a bit of a hack due to the fact that we aren't
instantiating  regularly. Normally, instantiated connections
don't share their IKE SA, but here we need to share it. So there
is some juggling happening.

We also missed copying the trailing NUL from the whack string into
the chunk_t sec_label.

It fixes IKEv1 to send the narrowed label, instead of the connection
label.

Add sec_label to the pending and oppo bundle structs so we can pick
it up in the outI1 functions.

The outI1 functions now all copy the ACQUIREd label into the state
for later use with IKE payloads and XFRM messages.