pluto: do not call v2_expire_unused_ike_sa() when we are handling EXPIRE event
When our last Child SA received a kernel EXPIRE message, when we would
delete it, we could also delete the IKE SA because it was "unused".

check specifically for st_kernel_sa_expired and if detected, skip
deleting a childless IKE SA.

Probably, we should just always let the IKE SA linger.