Remove automatic forced routing.
When transport mode is used, this causes forced routes to
pre-nat ip-addresses. Because those addresses are not really
used with transport mode, routes are never removed.

So while this automatic route might help with some cases where
route to client network is not there it can cause problems with
routing table with l2tp-over-IPsec or other vpn types using transport mode.