pluto: re-instate sending IPsec labeled on non-acquired initiate
If a connection is started using --up, without a packet trigger,
we have no ACQUIREd policy label to use. Use the configured
"catch all" label in that case. This is required so both ends know
labeled IPsec is required for the connection. If not sent, this
connection cannot be distinguished from a true labeled IPsec
connection (and refused if needed)

The actual policy label on the IPsec SA is never used. The installed
policy will cause ACQUIRES which will trigger another IPsec SA for
the specific subset of policy label (which is matched with polmatch)