x509: sort DNS pubkeys before adding them
Sort using the raw pubkey payload (which is a pretty arbitrary
choice).

This way, the pubkey list is more stable, while still having the
latest additions at the front.

The alternative is to keep the pubkey list sorted (just sorting keys
when they are listed isn't sufficient).  It turns out that the list
order leaks into the strangest of places, for instance when logging
the keys when authentication fails.