libswan/ttoaddress.c: attempt to fix a crash in ttoaddress_dns
Cagney observed a crash in the version of ttoaddress_dns that uses getaddrinfo.
It looks as if freeaddrinfo is trying to free something that wasn't malloced.
The best explanation is that getaddrinfo does NOT set *res if some failure
is recognized (this is not documented).

This commit initialized res to NULL.  That should work if freeaddrinfo
does or does not set *res on failure.