crypto/logging: return diag_t, instead of STF_STATUS, from *authsig_and_log*()
If the function succeeds, it logs, if it fails it returns a diag_t
(which could only be one line).

For IKEv2, this puts the one line diagnostic, in the hands of the
state transition function (and oh so close, and yet so far, from
complete_v2_state_transition()).

Also update some comments based on Coverity and DHR's feedback.