ikev2: when initiating a delete exchange, delete larval children during send
Normally a delete child response includes the responder->initiator SPI
and that is used to identify and delete the child.  This way the
responder gets to shutdown the responder->initiator SA before the
initiator.

However, for a larval child (where it is the initiator rejecting the
response), the responder->initiator SPI isn't known and the
corresponding SA isn't real.  Hence, just delete the SA.