ikev2: drop "auth=IKEv2" from the IKE SA's cryptosuite log line
State what is; not what once was.

Trying to authenticate as part of establishing a protected channel
was one of those good idea found in IKEv1.  Reminding everyone of
this past mistake in an IKEv2 log line is just confusing (and now
that the code uses a jambuf instead of a monolythic snprintf()
fixing it is easy).