connections: drop end.is_port_wildcard, use protoport.has_port_wildcard
Also drop the IKEv1 quick mode line that cleared .is_port_wildcard
after updating the client's port.

Have to wonder if .has_port_wildcard and/or the narrowing bit are in
the wrong place.  For instance, conbine the two into a single bit-set
that can be passed to the narrowing code.