kernel: in raw_policy(OP) replace cur_spi with what_about_inbound
This replaces the logic:

  (op == KP_DELETE_INBOUND ||
  (op == KP_DELETE_OUTBOUND && ntohl(cur_spi) == SPI_HOLD));

which was determining when to ignore updates that don't match a policy
(it was too weak and supressed real errors; given pluto can track when
it added a policy it can determine when it needs deleting).

Instead callers pass in WHAT_ABOUT_INBOUND which specifies when an
update should work, and when the update is dubious and should be
eliminated.  For instance, when a Child SA exchange fails, there's no
point trying to delete incoming policy that was never added.