kernel: do not delete a hard expired sa from kernel
Do not send a delete request for a (hard) expired xfrm SA, because the
krenel already deleted this SA.
A hard expired SA is already removed by the kernel, before the pluto
recieves a XFRM_MSG_EXPIRE message.

This probably happens rarely in real life, hwoever, in some usecases,
especially in test setup this will show up more often. A test setup
has low ipsec-max-bytes value, say 2KiB.
When the traffic is symetric, both SAs will expire one after the other.
While procssing the first hard expire message for one direction the second
SA has alreday expired in the kernel, and netlink message is queued for the pluto
to process.
It is likely a corner case in real life, because traffic is not likely
symetric and margin is big in GiB. While in a test setup
if you set limit as 2KiB and use 1500 byte packet several corner case
wil trigger at once. And appear to cerate chaos.