enumcheck: placate Coverity Scan; eliminate one heap use
Coverity didn't see that a second call to strchr couldn't yield NULL
if an earlier call with the same arguments didn't yield NULL.
Eliminating the second call should fix it.

The code duplicated a string into the heap in order to truncate it.
We eliminated the need for a copy by using length-bounded string
operations.