pluto: ikev1-policy=reject is also not supported without USE_IKEv1
As it is the IKEv1 code that generates the ICMP response, and that
code is not compiled in, all pluto can do is drop the packet. The
code already does this, but setting ikev1-policy=reject caused us
to not dbg/log this.

Thanks to Daniel Kahn Gillmor for spotting this.

Resolves: https://github.com/libreswan/libreswan/pull/606