pluto: ikev1_quick.c: paper over a bug in fc_try()
Coverity Scan found an uninitialized variable reference in fc_try.
I have initialize those variables, but possibly to an incorrect value.
The bug was introduced in 605c8010007.

I made parallel code in ikev2_ts.c slightly clearer.
It does not have the bug because the uninitialized variables are never used.