ikev1: don't zap RSASIG+PSK when that was proposed by peer
Given a proposal with both RSASIG+PSK, authby (nee policy) was
being cleared of any authentication bits.  Logic in find host-pair
then interpreted that as accept any.

Now the code matches like for like.  This means that an IKEv1
Main Mode proposal containing PSK is dropped when the
only connection option is authby=secret (which gives the fuzzy
tests grief).