connections: merge .has_lease and .has_internal_address
Call it .has_lease since it is probably what comes to mind
when CP is discussed.

Note that this fixes:

-000 "..."[1]:     oriented; my_ip=unset; their_ip=unset; ...
+000 "..."[1]:     oriented; my_ip=unset; their_ip=192.0.2.1; ...

"my_ip" and "their_ip" show the sourceip for each end (when unset
it is derived).  After the change, and just like the peer's output,
the peer's lease is shown.

ref #924