kernel: make assign_holdpass() totally responsible for widening
i.e., dump all the code related to widening a shunt into this
function and call it unconditionally.

Note that widening, here, refers to replacing the acquire's
kernel policy with that of the connection's SPD.  Except there
may be no acquire kernel policy and/or it may not need changing.
Oh and adding and then removing a bare_shunt entry is pointless.