sec_label: make a labeled_parent() a CK_INSTANCE, not a CK_TEMPLATE
sec_labels connections are instantiated:
  template -> parent (IKE) -> child (IPsec)
The "parent" connection, which is both kind of template and kind
of instance like, was being instantiated as a CK_TEMPLATE.  This
meant that the template routing code had to handle some pretty
weird sec_label specific transitions.   Changing it to CK_INSTANCE
lets that code be dumped.