confread: introduce ppk-ids connection option
- the option allows to define which PPK secrets in .secrets file
to look for using corresponding PPK_IDs. The option receives a
string containing comma-separated list of characters. PPK_ID can
contain whitespaces, they are not stripped.
- change accordingly the PPK code logic. If ppk-ids: connection
option is not specified, the code works as it has been working,
looking for first matching secret. Otherwise, one of the specified
PPK_IDs in ppk-ids option should be found in .secrets

This is a preparation for draft-smyslov-ipsecme-ikev2-qr-alt-07,
which allows for sending multiple PPK_IDs (each in it's own NOTIFY)