netlink_attrib: guard memcpy from invalid pointers
A common assumption is that memcpy address arguments need not be valid
if the length is zero.  The C standard says this is wrong.
In the case of this memcpy call, the source pointer might be NULL
if the length is zero.
We now skip the memcpy if the length is 0.

Found by lclint.