pluto: Add support for nic-offload=packet
Based on patches by Raed Salem <raeds@nvidia.com>
Requires Linux kernel 6.3+

This offload extends the current crypto offload where in addition
to the crypto operations, once can now offload the entire cleartext
packet to be encapsulated and encrypted by hardware offload.

This includes managing the IPsec SA policy in the offload hardware,
and requires additional IKE holes for the hardware to ensure IKE
packets are not required to be ESP encrypted.