connections: addref/delref when initiating connections
the model here is:

   connection = find_and_newref_connection_to_initiate()
   initiate connection; which takes a reference
   delref(connection)

where find_and_newref_connection_to_initiate()

- addref's permanent connections
- instantiates template connections

With IKEv2 the nested initiate functions follow the model closely
IKEv1 is a bit tangled.

Note that in addition to the initiate code, this also happens when
instantiating connections when finding a better connection.

The only got-ya is labeled-templates.  Because the labeled-parent
is permanent like it's reference isn't released.