initiate: honour .viable_parent
When looking for an IKE SA to initiate a new Child SA check that
an established candidate is still viable (as in not being deleted).

Since .viable_parent started in the IKEv2 code, IKEv1 has never
had reason to check this.  IKEv2 does check it but only after commiting
to an IKE SA.