ikev1: move clear_retransmits() in submit_task() to STF_SUSPEND
close #657; note #1339

Add the note on why it it is still broken:

XXX: Clearing retransmits here is wrong (it is a
slight improvement on submit_task()).

Retransmits should only be cleared after the
integrity of the packet has been proven and here
that is likely not the case.  For instance, the
exchange is suspended while the DH needed to prove
integrity is computed.

A better location might be in STF_v1N, assuming the
packet's integrity was verified.