redirect: use revival and routing code to initiate redirects
This way the revival and redirect code can't get into a fight,
fixing #1396 and #1319.

- when a redirect arrives, save the redirect details in connection
  .redirect and then trigger an ike family delete (return
  STF_OK_INITIATOR_DELETE_IKE for IKE_SA_INIT and IKE_AUTH, schedule
  DISCARD event for established).

- the routing code sees the +UP connection going down and schedules an
  immediate revival (while also updating kernel state and policy as
  required)

- the revival code sees that it is for a redirect, and acts
  accordingly

When the redirect limit is exceeded, the step saving the redirect
details is skipped and instead revival is forced to 300s per RFC
(#1401).  Consequently a normal, if much delayed, revival is
scheduled.  A failing redirect is the same, but with a more sensible
delay.