pending: drop EVENT_PENDING_PHASE2 reviving lost connection
In v5.x the pending list is directly attached to the IKE SA, so when
that dies they are always revived.  In v4.x, while they were attached
to the host-pair, an IKE SA death caused the list to be scanned and
pending re-initiated.

Here's the original description:

  Look for phase2s that were waiting for a phase 1.  If the time that
  we have been pending exceeds a DPD timeout that was set, then we
  call the dpd_timeout() on this state, which hopefully kills this
  pending state.

That last bit - hopefully - really inspires confidence :-)