testing: swan-prep fixes
swan-prep shouldn't really be called on nic. It is used to setup DNS on
nic though. So the swan-prep check for eth0 fails when using namespaces
on nic.

Also, the nsd keygen service no longer exists (a socket is used instead
of a TLS connection to localhost)

Unfixed still is the missing nsd.conf when using namespaces.