ikev2: disable non-MOBIKE NAT endpoint updates
The idea is for an IKE SA that receives an authenticated packet
from it's NATed peer, but with a new address, should update
the peer's address as, presumably, NAT updated things.

The feature was only haf implemented:
  - IKE SA's endpoint was updated
  - IPsec kernel state/policy endpoints was left unchanged
The result was an IKE SA thinking all was good when no actual traffic could
flow.

see:
  IKEv2 liveness does not work with IP change #1492
    where @wofferl explains the problem
  implement IKEv2's non-MOBIKE NAT port/address updates #1529