pluto: remove nic-offload=auto
It is complicated to make this work as we need to load the policy
matching for crypto or packet offload before we know if packet
offload is supported for the negotiated parameters of the IPsec SA.

For now, only allow "packet" or "crypto". Don't attempt any fallbacks
ourselves. On Linux, the kernel provides crypto to none fallback
for AEADs (or at least for AES-GCM)