state: save negotiated mode (transport, tunnel, ...) in .st_kernel_mode
This replaces .st_{esp,ah,ipcomp}.attrs.mode

For IKEv1 the mode is negotiated as part of the proposal (which
explains where this came from).  This is tweaked to store the final
decision in .st_kernel_mode.

For IKEv2, the mode is negotiated using notifies (it never had
anything to do with proposals) so the decision is stored directly in
.st_kernel_mode.

The field attrs.mode is dropped and any code hunting and pecking for
its value deleted.