libipsecconf: change also= to recursively expand inline
Re-implement also= so that, when building the connection, the entries
are expanded as they are encountered.  Also drop arbitrary also=
limit.

This changes the behaviour of also= so that it matches typical #include
and macro expansion.  For instance, given:

  conn a
    also=b
    also=d
    x=a
  conn b
    also=c
    x=b
  conn c
    x=c
  conn d
    x=d

which previously was expanded as:

  a,b,d,c -- x=a

is now now expanded inline vis:

  a,b,c,d -- x=c

i.e., also= entries are processed as they are encountered (and not
appended to a queue and processed later)

Update documentation.