host-pair: search connections old-to-new
This brings the code into line with check_orientations().  That
function was changed to OLD2NEW so that it didn't flip the order of
aliased connections when adding an interface.

This "fixes" delete-sa-08-rhbz1311360.
This "breaks" whack-delete-01-ikev1-child-1then2.

The bottom line is that both tests are relying on unspecified
behaviour.