whack: define struct whack_basic; containing .whack_{shutdown,status}
Move all the non-basic fields (.whack_briefstatus et.al.) to beyond
the struct so that, hopefully, it's clear what is basic ...
Size the struct so it covers the over-the-years locations of
.whack_shutdown.

Split --shutdown --leave-state and --shutdown so that .whack_shutdown
does not need to refer to .whack_leave_state.

Move code setting and checking .magic, and checking size of
read buffer, into pickle.c.  Only allow whack_magic() and/or
WHACK_BASIC_MAGIC.

See #1709