re-merge: xfrmi: fix adding xfrm interface ip address with ipv6
(the original merge was botched doe to a code clash)

The code actual checks for IFA_LOCAL and IFA_LABEL when searching for
an existing ip address.  This doesn't work with ipv6, because the
kernel only returns IFA_ADDRESS for ipv6.  Since the xfrm interface is
always a broadcast interface, we can always use IFA_ADDRESS for both
ipv4 and ipv6.  IFA_LABEL is not necessary because you can filter on
the interface index.

ref #1741 fixes for xfrm interface-ip and sourceip
fix #1765 test ikev2-xfrmi-15-interface-ip-blocked failing, bad merge?
ref #1769 xfrmi support found existing ipsec1@^E xfrm if_id 0x1 - note the ^E

Signed-off-by: Andrew Cagney <cagney@gnu.org>